Project/com/youngho/spring/spring-demo-project/src/main/java/com/youngho/springdemoproject/repository/MemoryMemberRepository.java
package com.youngho.springdemoproject.repository;

import com.youngho.springdemoproject.domain.Member;
import org.springframework.stereotype.Repository;

import java.util.HashMap;
import java.util.List;
import java.util.Optional;

import java.util.*;

/*
* 동시성 문제를 고려하지 않음, 실무에서는 ConcurrentHashMap, AtomicLong 사용 고려
* */
@Repository
public class MemoryMemberRepository implements MemberRepository {

    private static final Map<Long, Member> store = new HashMap<>();
    private static long sequence = 0L;

    @Override
    public Member save(Member member) {
        member.setId(++sequence);
        store.put(member.getId(), member);      // 해시맵에 (key, value)의 값으로 저장됨
        return member;
    }

    @Override
    public Optional<Member> findById(Long id) {
        return Optional.ofNullable(store.get(id));          /* id가 없으면 return Optional, 그렇지 않으면 return id */
    }

    @Override
    public Optional<Member> findByName(String name) {
        return store.values().stream()
                .filter(member -> member.getName().equals(name))
                .findAny();
    }

    @Override
    public List<Member> findAll() {
        return new ArrayList<>(store.values());
    }

    public void clearStore() {
        store.clear();
    }
}
